chatbot/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
openai-whisper==20230314
torch>=1.12.1
